Moved all helper methods to a different file
