+++ tests/backup/backup_test.go
@@ -4,10 +4,6 @@ import (
 	"context"
 	"fmt"
 	"math/rand"
-	"os"
-	"os/exec"
-	"path"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -21,7 +17,6 @@ import (
 	"github.com/portworx/sched-ops/k8s/core"
 	storage "github.com/portworx/sched-ops/k8s/storage"
 	"github.com/portworx/sched-ops/task"
-	driver_api "github.com/portworx/torpedo/drivers/api"
 	"github.com/portworx/torpedo/drivers/backup"
 	"github.com/portworx/torpedo/drivers/backup/portworx"
 	"github.com/portworx/torpedo/drivers/node"
@@ -36,33 +31,6 @@ import (
 
 	. "github.com/portworx/torpedo/tests"
 )
-
-type BackupAccess int32
-
-const (
-	ViewOnlyAccess BackupAccess = 1
-	RestoreAccess               = 2
-	FullAccess                  = 3
-)
-
-type userRoleAccess struct {
-	user     string
-	roles    backup.PxBackupRole
-	accesses BackupAccess
-	context  context.Context
-}
-
-type userAccessContext struct {
-	user     string
-	accesses BackupAccess
-	context  context.Context
-}
-
-var backupAccessKeyValue = map[BackupAccess]string{
-	1: "ViewOnlyAccess",
-	2: "RestoreAccess",
-	3: "FullAccess",
-}
 
 // This testcase verifies if the backup pods are in Ready state or not
 var _ = Describe("{BackupClusterVerification}", func() {
@@ -386,15 +354,18 @@ var _ = Describe("{DuplicateSharedBackup}", func() {
 		log.FailOnError(err, "Backup [%s] could not be deleted with delete response %s", backupName, backupDeleteResponse)
 
 		log.Infof("Deleting registered clusters for admin context")
-		DeleteCluster(SourceClusterName, orgID, ctx)
-		DeleteCluster(destinationClusterName, orgID, ctx)
+		err = DeleteCluster(SourceClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 
 		log.Infof("Cleaning up backup location - %s", BackupLocationName)
 		DeleteBackupLocation(BackupLocationName, backupLocationUID, orgID)
 
 		log.Infof("Cleaning cloud credential")
 		time.Sleep(time.Minute * 3)
-		DeleteCloudCredential(credName, orgID, cloudCredUID)
+		err = DeleteCloudCredential(credName, orgID, cloudCredUID)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", credName))
 
 	})
 
@@ -1151,13 +1122,17 @@ var _ = Describe("{ShareBackupWithUsersAndGroups}", func() {
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 
 		log.Infof("Deleting registered clusters for admin context")
-		DeleteCluster(SourceClusterName, orgID, ctx)
-		DeleteCluster(destinationClusterName, orgID, ctx)
+		err = DeleteCluster(SourceClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 
 		log.Infof("Deleting registered clusters for non-admin context")
 		for _, ctxNonAdmin := range userContexts {
-			DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
-			DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			err = DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+			err = DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 		}
 
 		backupDriver := Inst().Backup
@@ -1176,7 +1151,8 @@ var _ = Describe("{ShareBackupWithUsersAndGroups}", func() {
 		log.Infof("Cleaning cloud credential")
 		//TODO: Eliminate time.Sleep
 		time.Sleep(time.Minute * 3)
-		DeleteCloudCredential(credName, orgID, cloudCredUID)
+		err = DeleteCloudCredential(credName, orgID, cloudCredUID)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", credName))
 	})
 })
 
@@ -1509,13 +1485,17 @@ var _ = Describe("{ShareLargeNumberOfBackupsWithLargeNumberOfUsers}", func() {
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 
 		log.Infof("Deleting registered clusters for admin context")
-		DeleteCluster(SourceClusterName, orgID, ctx)
-		DeleteCluster(destinationClusterName, orgID, ctx)
+		err = DeleteCluster(SourceClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 
 		log.Infof("Deleting registered clusters for non-admin context")
 		for _, ctxNonAdmin := range userContexts {
-			DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
-			DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			err = DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+			err = DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 		}
 
 		backupDriver := Inst().Backup
@@ -1533,7 +1513,8 @@ var _ = Describe("{ShareLargeNumberOfBackupsWithLargeNumberOfUsers}", func() {
 		log.Infof("Cleaning cloud credential")
 		//TODO: Eliminate time.Sleep
 		time.Sleep(time.Minute * 3)
-		DeleteCloudCredential(credName, orgID, cloudCredUID)
+		err = DeleteCloudCredential(credName, orgID, cloudCredUID)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", credName))
 	})
 })
 
@@ -2141,7 +2122,9 @@ var _ = Describe("{ShareBackupAndEdit}", func() {
 			backupDriver := Inst().Backup
 			backupUID, err := backupDriver.GetBackupUID(ctx, backupNames[0], orgID)
 			log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupNames[0])
-			UpdateBackup(backupNames[0], backupUID, orgID, newCredName, newCloudCredUID, ctx)
+			status, err := UpdateBackup(backupNames[0], backupUID, orgID, newCredName, newCloudCredUID, ctx)
+			dash.VerifyFatal(err, nil, fmt.Sprintf("Updating backup %s with new cred %v", backupNames[0], newCredName))
+			log.Infof("The status after updating backup %s with new cred %v is %v", backupNames[0], newCredName, status)
 
 			// Get user context
 			ctxNonAdmin, err := backup.GetNonAdminCtx(users[0], "Password1")
@@ -2194,7 +2177,9 @@ var _ = Describe("{ShareBackupAndEdit}", func() {
 
 			//update the backup with another cred
 			log.InfoD("Update the backup with another cred")
-			UpdateBackup(backupNames[0], backupUID, orgID, credName, cloudCredUID, ctxNonAdmin)
+			status, err := UpdateBackup(backupNames[0], backupUID, orgID, credName, cloudCredUID, ctxNonAdmin)
+			dash.VerifyFatal(err, nil, fmt.Sprintf("Updating backup %s with new cred %v", backupNames[0], credName))
+			log.Infof("The status after updating backup %s with new cred %v is %v", backupNames[0], credName, status)
 
 			// Start Restore
 			restoreName := fmt.Sprintf("%s-%v", RestoreNamePrefix, time.Now().Unix())
@@ -2239,13 +2224,17 @@ var _ = Describe("{ShareBackupAndEdit}", func() {
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 
 		log.Infof("Deleting registered clusters for admin context")
-		DeleteCluster(SourceClusterName, orgID, ctx)
-		DeleteCluster(destinationClusterName, orgID, ctx)
+		err = DeleteCluster(SourceClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 
 		log.Infof("Deleting registered clusters for non-admin context")
 		for _, ctxNonAdmin := range userContexts {
-			DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
-			DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			err = DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+			err = DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 		}
 
 		log.Infof("Cleaning up backup location - %s", BackupLocationName)
@@ -2254,8 +2243,10 @@ var _ = Describe("{ShareBackupAndEdit}", func() {
 		log.Infof("Cleaning cloud credential")
 		//TODO: Eliminate time.Sleep
 		time.Sleep(time.Minute * 3)
-		DeleteCloudCredential(credName, orgID, cloudCredUID)
-		DeleteCloudCredential(newCredName, orgID, newCloudCredUID)
+		err = DeleteCloudCredential(credName, orgID, cloudCredUID)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", credName))
+		err = DeleteCloudCredential(newCredName, orgID, newCloudCredUID)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", newCredName))
 	})
 })
 
@@ -2454,13 +2445,17 @@ var _ = Describe("{SharedBackupDelete}", func() {
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 
 		log.Infof("Deleting registered clusters for admin context")
-		DeleteCluster(SourceClusterName, orgID, ctx)
-		DeleteCluster(destinationClusterName, orgID, ctx)
+		err = DeleteCluster(SourceClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 
 		log.Infof("Deleting registered clusters for non-admin context")
 		for _, ctxNonAdmin := range userContexts {
-			DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
-			DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			err = DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+			err = DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 		}
 
 		log.Infof("Cleaning up backup location - %s", BackupLocationName)
@@ -2469,7 +2464,8 @@ var _ = Describe("{SharedBackupDelete}", func() {
 		log.Infof("Cleaning cloud credential")
 		//TODO: Eliminate time.Sleep
 		time.Sleep(time.Minute * 3)
-		DeleteCloudCredential(credName, orgID, cloudCredUID)
+		err = DeleteCloudCredential(credName, orgID, cloudCredUID)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", credName))
 	})
 })
 
@@ -2640,14 +2636,17 @@ var _ = Describe("{BackupAlternatingBetweenLockedAndUnlockedBuckets}", func() {
 		// Need sleep as it takes some time for
 		time.Sleep(time.Minute * 1)
 		for CloudCredUID, CredName := range CloudCredUIDMap {
-			DeleteCloudCredential(CredName, orgID, CloudCredUID)
+			err := DeleteCloudCredential(CredName, orgID, CloudCredUID)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", CredName))
 		}
 		ctx, err := backup.GetAdminCtxFromSecret()
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 
 		log.Infof("Deleting registered clusters for admin context")
-		DeleteCluster(SourceClusterName, orgID, ctx)
-		DeleteCluster(destinationClusterName, orgID, ctx)
+		err = DeleteCluster(SourceClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 	})
 })
 
@@ -3021,8 +3020,10 @@ var _ = Describe("{ShareBackupsAndClusterWithUser}", func() {
 		}
 		log.FailOnError(err, "Error in deleting user")
 		log.Infof("Deleting registered clusters for non-admin context")
-		DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
-		DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+		err = DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 		DeleteCloudAccounts(backupLocationMap, cloudCredName, cloudCredUID, ctx)
 	})
 })
@@ -3377,13 +3378,17 @@ var _ = Describe("{DeleteSharedBackup}", func() {
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 
 		log.Infof("Deleting registered clusters for admin context")
-		DeleteCluster(SourceClusterName, orgID, ctx)
-		DeleteCluster(destinationClusterName, orgID, ctx)
+		err = DeleteCluster(SourceClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+		err = DeleteCluster(destinationClusterName, orgID, ctx)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 
 		log.Infof("Deleting registered clusters for non-admin context")
 		for _, ctxNonAdmin := range userContexts {
-			DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
-			DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			err = DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+			err = DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 		}
 
 		log.Infof("Cleaning up backup location - %s", BackupLocationName)
@@ -3392,7 +3397,8 @@ var _ = Describe("{DeleteSharedBackup}", func() {
 		log.Infof("Cleaning cloud credential")
 		//TODO: Eliminate time.Sleep
 		time.Sleep(time.Minute * 3)
-		DeleteCloudCredential(credName, orgID, cloudCredUID)
+		err = DeleteCloudCredential(credName, orgID, cloudCredUID)
+		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", credName))
 
 	})
 
@@ -3508,8 +3514,9 @@ var _ = Describe("{BackupRestartPX}", func() {
 				preRuleUid, _ := Inst().Backup.GetRuleUid(orgID, ctx, preRuleNameList[0])
 				postRuleUid, _ := Inst().Backup.GetRuleUid(orgID, ctx, postRuleNameList[0])
 				backupName := fmt.Sprintf("%s-%s-%s", BackupNamePrefix, namespace, backupLocationName)
-				CreateBackupWithoutCheck(backupName, SourceClusterName, backupLocation, backupLocationUID, []string{namespace},
+				err = CreateBackupWithoutCheck(backupName, SourceClusterName, backupLocation, backupLocationUID, []string{namespace},
 					labelSelectors, orgID, clusterUid, preRuleNameList[0], preRuleUid, postRuleNameList[0], postRuleUid, ctx)
+				dash.VerifyFatal(err, nil, fmt.Sprintf("Creating backup %s", backupName))
 			}
 		})
 
@@ -4043,7 +4050,8 @@ var _ = Describe("{BackupRestoreOverPeriod}", func() {
 				}
 			})
 			Step("teardown backup/restore objects", func() {
-				TearDownBackupRestoreSpecific(successfulBackupNames, successfulRestoreNames)
+				err := TearDownBackupRestoreSpecific(successfulBackupNames, successfulRestoreNames)
+				dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting backup %v and restore %v", successfulBackupNames, successfulRestoreNames))
 			})
 		}
 		Step("report statistics", func() {
@@ -4310,7 +4318,8 @@ var _ = Describe("{BackupRestoreOverPeriodSimultaneous}", func() {
 			}
 		})
 		Step("teardown backup/restore objects", func() {
-			TearDownBackupRestoreSpecific(successfulBackupNames, successfulRestoreNames)
+			err := TearDownBackupRestoreSpecific(successfulBackupNames, successfulRestoreNames)
+			dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting backup %v and restore %v", successfulBackupNames, successfulRestoreNames))
 		})
 		Step("report statistics", func() {
 			log.Infof("%d/%d backups succeeded.", successfulBackups, numBackups)
@@ -4440,8 +4449,9 @@ var _ = Describe("{KillStorkWithBackupsAndRestoresInProgress}", func() {
 				preRuleUid, _ := Inst().Backup.GetRuleUid(orgID, ctx, preRuleNameList[0])
 				postRuleUid, _ := Inst().Backup.GetRuleUid(orgID, ctx, postRuleNameList[0])
 				backupName := fmt.Sprintf("%s-%s-%v", BackupNamePrefix, namespace, time.Now().Unix())
-				CreateBackupWithoutCheck(backupName, SourceClusterName, backupLocation, backupLocationUID, []string{namespace},
+				err = CreateBackupWithoutCheck(backupName, SourceClusterName, backupLocation, backupLocationUID, []string{namespace},
 					labelSelectors, orgID, clusterUid, preRuleNameList[0], preRuleUid, postRuleNameList[0], postRuleUid, ctx)
+				dash.VerifyFatal(err, nil, fmt.Sprintf("Creating backup %s", backupName))
 				backupNames = append(backupNames, backupName)
 			}
 		})
@@ -4660,7 +4670,8 @@ var _ = Describe("{BackupLocationWithEncryptionKey}", func() {
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 		err = DeleteRestore(restoreName, orgID, ctx)
 		dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting Restore %s", restoreName))
-		backupUID := getBackupUID(backupName, orgID)
+		backupUID, err := getBackupUID(backupName, orgID)
+		dash.VerifyFatal(err, nil, fmt.Sprintf("Getting backuip UID for backup %s", backupName))
 		_, err = DeleteBackup(backupName, backupUID, orgID, ctx)
 		dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting backup %s", backupName))
 		DeleteCloudAccounts(backupLocationMap, cloudCredName, CloudCredUID, ctx)
@@ -5161,7 +5172,8 @@ var _ = Describe("{RestoreEncryptedAndNonEncryptedBackups}", func() {
 		ctx, err = backup.GetAdminCtxFromSecret()
 		log.FailOnError(err, "Fetching px-central-admin ctx")
 		for _, backupName := range backupNames {
-			backupUID := getBackupUID(backupName, orgID)
+			backupUID, err := getBackupUID(backupName, orgID)
+			dash.VerifyFatal(err, nil, fmt.Sprintf("Getting backuip UID for backup %s", backupName))
 			_, err = DeleteBackup(backupName, backupUID, orgID, ctx)
 			dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting backup %s", backupName))
 		}
@@ -5296,8 +5308,9 @@ var _ = Describe("{CustomResourceBackupAndRestore}", func() {
 			dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting Restore %s", restore))
 		}
 		for _, backupName := range backupNames {
-			backupUID := getBackupUID(backupName, orgID)
-			_, err := DeleteBackup(backupName, backupUID, orgID, ctx)
+			backupUID, err := getBackupUID(backupName, orgID)
+			dash.VerifyFatal(err, nil, fmt.Sprintf("Getting backuip UID for backup %s", backupName))
+			_, err = DeleteBackup(backupName, backupUID, orgID, ctx)
 			dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting backup - %s", backupName))
 		}
 		log.InfoD("Deleting backup location, cloud creds and clusters")
@@ -5444,8 +5457,9 @@ var _ = Describe("{ReplicaChangeWhileRestore}", func() {
 			err := Inst().S.Destroy(contexts[i], opts)
 			dash.VerifySafely(err, nil, fmt.Sprintf("Verify destroying app %s", taskName))
 		}
-		backupUID := getBackupUID(backupName, orgID)
-		_, err := DeleteBackup(backupName, backupUID, orgID, ctx)
+		backupUID, err := getBackupUID(backupName, orgID)
+		dash.VerifyFatal(err, nil, fmt.Sprintf("Getting backuip UID for backup %s", backupName))
+		_, err = DeleteBackup(backupName, backupUID, orgID, ctx)
 		dash.VerifySafely(err, nil, fmt.Sprintf("Deleting backup [%s]", backupName))
 		err = DeleteRestore(restoreName, orgID, ctx)
 		dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting Restore [%s]", restoreName))
@@ -5878,8 +5892,10 @@ var _ = Describe("{ViewOnlyFullBackupRestoreIncrementalBackup}", func() {
 
 		log.Infof("Deleting registered clusters for non-admin context")
 		for _, ctxNonAdmin := range userContexts {
-			DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
-			DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			err = DeleteCluster(SourceClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
+			err = DeleteCluster(destinationClusterName, orgID, ctxNonAdmin)
+			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
 		}
 
 		log.Infof("Cleaning up user")
@@ -7410,1168 +7426,3 @@ var _ = Describe("{ScheduleBackupCreationAllNS}", func() {
 		DeleteCloudAccounts(backupLocationMap, cloudAccountName, cloudCredUID, ctx)
 	})
 })
-
-// createS3BackupLocation creates backup location
-func createGkeBackupLocation(name string, cloudCred string, orgID string) {
-	Step(fmt.Sprintf("Create GKE backup location [%s] in org [%s]", name, orgID), func() {
-		// TODO(stgleb): Implement this
-	})
-}
-
-// CreateProviderClusterObject creates cluster for each cluster per each cloud provider
-func CreateProviderClusterObject(provider string, kubeconfigList []string, cloudCred, orgID string, ctx context.Context) {
-	Step(fmt.Sprintf("Create cluster [%s-%s] in org [%s]",
-		clusterName, provider, orgID), func() {
-		kubeconfigPath, err := getProviderClusterConfigPath(provider, kubeconfigList)
-		log.FailOnError(err, "Fetching kubeconfig path for source cluster")
-		CreateCluster(fmt.Sprintf("%s-%s", clusterName, provider),
-			kubeconfigPath, orgID, "", "", ctx)
-	})
-}
-
-func getProviders() []string {
-	providersStr := os.Getenv("PROVIDERS")
-	return strings.Split(providersStr, ",")
-}
-
-// getPXNamespace fetches px namespace from env else sends backup kube-system
-func getPXNamespace() string {
-	namespace := os.Getenv("PX_NAMESPACE")
-	if namespace != "" {
-		return namespace
-	}
-	return storkDeploymentNamespace
-}
-
-func getProviderClusterConfigPath(provider string, kubeconfigs []string) (string, error) {
-	log.Infof("Get kubeconfigPath from list %v and provider %s",
-		kubeconfigs, provider)
-	for _, kubeconfigPath := range kubeconfigs {
-		if strings.Contains(provider, kubeconfigPath) {
-			fullPath := path.Join(KubeconfigDirectory, kubeconfigPath)
-			return fullPath, nil
-		}
-	}
-
-	return "nil", fmt.Errorf("kubeconfigPath not found for provider %s in kubeconfigPath list %v",
-		provider, kubeconfigs)
-}
-
-// CreateBackup creates backup
-func CreateBackup(backupName string, clusterName string, bLocation string, bLocationUID string,
-	namespaces []string, labelSelectors map[string]string, orgID string, uid string, preRuleName string,
-	preRuleUid string, postRuleName string, postRuleUid string, ctx context.Context) error {
-
-	var bkpUid string
-	backupDriver := Inst().Backup
-	bkpCreateRequest := &api.BackupCreateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  backupName,
-			OrgId: orgID,
-		},
-		BackupLocationRef: &api.ObjectRef{
-			Name: bLocation,
-			Uid:  bLocationUID,
-		},
-		Cluster:        clusterName,
-		Namespaces:     namespaces,
-		LabelSelectors: labelSelectors,
-		ClusterRef: &api.ObjectRef{
-			Name: clusterName,
-			Uid:  uid,
-		},
-		PreExecRuleRef: &api.ObjectRef{
-			Name: preRuleName,
-			Uid:  preRuleUid,
-		},
-		PostExecRuleRef: &api.ObjectRef{
-			Name: postRuleName,
-			Uid:  postRuleUid,
-		},
-	}
-	_, err := backupDriver.CreateBackup(ctx, bkpCreateRequest)
-	if err != nil {
-		return err
-	}
-	backupSuccessCheck := func() (interface{}, bool, error) {
-		bkpUid, err = backupDriver.GetBackupUID(ctx, backupName, orgID)
-		log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupName)
-		backupInspectRequest := &api.BackupInspectRequest{
-			Name:  backupName,
-			Uid:   bkpUid,
-			OrgId: orgID,
-		}
-		resp, err := backupDriver.InspectBackup(ctx, backupInspectRequest)
-		log.FailOnError(err, "Inspecting the backup taken with request:\n%v", backupInspectRequest)
-		actual := resp.GetBackup().GetStatus().Status
-		expected := api.BackupInfo_StatusInfo_Success
-		if actual != expected {
-			return "", true, fmt.Errorf("backup status for [%s] expected was [%s] but got [%s]", backupName, expected, actual)
-		}
-		return "", false, nil
-	}
-
-	task.DoRetryWithTimeout(backupSuccessCheck, maxWaitPeriodForBackupCompletionInMinutes*time.Minute, 30*time.Second)
-
-	bkpUid, err = backupDriver.GetBackupUID(ctx, backupName, orgID)
-	log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupName)
-	backupInspectRequest := &api.BackupInspectRequest{
-		Name:  backupName,
-		Uid:   bkpUid,
-		OrgId: orgID,
-	}
-	_, err = backupDriver.InspectBackup(ctx, backupInspectRequest)
-	if err != nil {
-		return err
-	}
-	log.Infof("Backup [%s] created successfully", backupName)
-	return nil
-}
-func UpdateBackup(backupName string, backupuid string, org_id string, cloudCred string, cloudCredUID string, ctx context.Context) {
-	backupDriver := Inst().Backup
-	bkpUpdateRequest := &api.BackupUpdateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  backupName,
-			OrgId: orgID,
-			Uid:   backupuid,
-		},
-		CloudCredential: cloudCred,
-		CloudCredentialRef: &api.ObjectRef{
-			Name: cloudCred,
-			Uid:  cloudCredUID,
-		},
-	}
-	_, err := backupDriver.UpdateBackup(ctx, bkpUpdateRequest)
-	log.FailOnError(err, "Failed to update backup with request -\n%v", bkpUpdateRequest)
-
-}
-
-// CreateBackup creates backup with custom resources
-func CreateBackupWithCustomResourceType(backupName string, clusterName string, bLocation string, bLocationUID string,
-	namespaces []string, labelSelectors map[string]string, orgID string, uid string, preRuleName string,
-	preRuleUid string, postRuleName string, postRuleUid string, resourceType []string, ctx context.Context) error {
-
-	var bkpUid string
-	backupDriver := Inst().Backup
-	bkpCreateRequest := &api.BackupCreateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  backupName,
-			OrgId: orgID,
-		},
-		BackupLocationRef: &api.ObjectRef{
-			Name: bLocation,
-			Uid:  bLocationUID,
-		},
-		Cluster:        clusterName,
-		Namespaces:     namespaces,
-		LabelSelectors: labelSelectors,
-		ClusterRef: &api.ObjectRef{
-			Name: clusterName,
-			Uid:  uid,
-		},
-		PreExecRuleRef: &api.ObjectRef{
-			Name: preRuleName,
-			Uid:  preRuleUid,
-		},
-		PostExecRuleRef: &api.ObjectRef{
-			Name: postRuleName,
-			Uid:  postRuleUid,
-		},
-		ResourceTypes: resourceType,
-	}
-	_, err := backupDriver.CreateBackup(ctx, bkpCreateRequest)
-	if err != nil {
-		return err
-	}
-	backupSuccessCheck := func() (interface{}, bool, error) {
-		bkpUid, err = backupDriver.GetBackupUID(ctx, backupName, orgID)
-		log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupName)
-		backupInspectRequest := &api.BackupInspectRequest{
-			Name:  backupName,
-			Uid:   bkpUid,
-			OrgId: orgID,
-		}
-		resp, err := backupDriver.InspectBackup(ctx, backupInspectRequest)
-		log.FailOnError(err, "Inspecting the backup taken with request:\n%v", backupInspectRequest)
-		actual := resp.GetBackup().GetStatus().Status
-		expected := api.BackupInfo_StatusInfo_Success
-		if actual != expected {
-			return "", true, fmt.Errorf("backup status for [%s] expected was [%s] but got [%s]", backupName, expected, actual)
-		}
-		return "", false, nil
-	}
-
-	task.DoRetryWithTimeout(backupSuccessCheck, 10*time.Minute, 30*time.Second)
-
-	bkpUid, err = backupDriver.GetBackupUID(ctx, backupName, orgID)
-	log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupName)
-	backupInspectRequest := &api.BackupInspectRequest{
-		Name:  backupName,
-		Uid:   bkpUid,
-		OrgId: orgID,
-	}
-	_, err = backupDriver.InspectBackup(ctx, backupInspectRequest)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// CreateScheduleBackup creates a scheduled backup
-func CreateScheduleBackup(backupName string, clusterName string, bLocation string, bLocationUID string,
-	namespaces []string, labelSelectors map[string]string, orgID string, preRuleName string,
-	preRuleUid string, postRuleName string, postRuleUid string, schPolicyName string, schPolicyUID string, ctx context.Context) (string, error) {
-	var bkpUid string
-	backupDriver := Inst().Backup
-	bkpSchCreateRequest := &api.BackupScheduleCreateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  backupName,
-			OrgId: orgID,
-		},
-		SchedulePolicyRef: &api.ObjectRef{
-			Name: schPolicyName,
-			Uid:  schPolicyUID,
-		},
-		BackupLocationRef: &api.ObjectRef{
-			Name: bLocation,
-			Uid:  bLocationUID,
-		},
-		SchedulePolicy: schPolicyName,
-		Cluster:        clusterName,
-		Namespaces:     namespaces,
-		LabelSelectors: labelSelectors,
-		PreExecRuleRef: &api.ObjectRef{
-			Name: preRuleName,
-			Uid:  preRuleUid,
-		},
-		PostExecRuleRef: &api.ObjectRef{
-			Name: postRuleName,
-			Uid:  postRuleUid,
-		},
-	}
-	_, err := backupDriver.CreateBackupSchedule(ctx, bkpSchCreateRequest)
-	if err != nil {
-		return "", err
-	}
-	time.Sleep(1 * time.Minute)
-	backupSuccessCheck := func() (interface{}, bool, error) {
-		bkpUid, err = backupDriver.GetBackupUID(ctx, backupName, orgID)
-		if err != nil {
-			return "", true, err
-		}
-		backupSchInspectRequest := &api.BackupScheduleInspectRequest{
-			Name:  backupName,
-			Uid:   bkpUid,
-			OrgId: orgID,
-		}
-
-		resp, err := backupDriver.InspectBackupSchedule(ctx, backupSchInspectRequest)
-		if err != nil {
-			return "", true, fmt.Errorf("Error in fetching inspect backup schedule response for %v", backupName)
-		}
-		expected := api.BackupScheduleInfo_StatusInfo_Success
-		actual := resp.GetBackupSchedule().GetBackupStatus()["interval"].GetStatus()[0].GetStatus()
-		if actual != expected {
-			return "", true, fmt.Errorf("backup status for [%s] expected was [%s] but got [%s]", backupName, expected, actual)
-		}
-		return fmt.Sprintf("actual [%v] is equal to expected [%v] string", actual, expected), false, nil
-	}
-
-	_, err = task.DoRetryWithTimeout(backupSuccessCheck, 10*time.Minute, 30*time.Second)
-	if err != nil {
-		return "", err
-	}
-	bkpUid, err = backupDriver.GetBackupUID(ctx, backupName, orgID)
-	if err != nil {
-		return "", err
-	}
-	backupSchInspectRequest := &api.BackupScheduleInspectRequest{
-		Name:  backupName,
-		Uid:   bkpUid,
-		OrgId: orgID,
-	}
-	resp, err := backupDriver.InspectBackupSchedule(ctx, backupSchInspectRequest)
-	log.InfoD("InspectBackupSchedule response - [%v]", resp)
-	if err != nil {
-		return "", err
-	}
-	schBackupName := resp.GetBackupSchedule().GetBackupStatus()["interval"].GetStatus()[0].GetBackupName()
-	log.InfoD("InspectBackupSchedule response - [%v]", schBackupName)
-	return schBackupName, nil
-}
-
-// CreateBackupWithoutCheck creates backup without waiting for success
-func CreateBackupWithoutCheck(backupName string, clusterName string, bLocation string, bLocationUID string,
-	namespaces []string, labelSelectors map[string]string, orgID string, uid string, preRuleName string,
-	preRuleUid string, postRuleName string, postRuleUid string, ctx context.Context) {
-
-	backupDriver := Inst().Backup
-	bkpCreateRequest := &api.BackupCreateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  backupName,
-			OrgId: orgID,
-		},
-		BackupLocationRef: &api.ObjectRef{
-			Name: bLocation,
-			Uid:  bLocationUID,
-		},
-		Cluster:        clusterName,
-		Namespaces:     namespaces,
-		LabelSelectors: labelSelectors,
-		ClusterRef: &api.ObjectRef{
-			Name: clusterName,
-			Uid:  uid,
-		},
-		PreExecRuleRef: &api.ObjectRef{
-			Name: preRuleName,
-			Uid:  preRuleUid,
-		},
-		PostExecRuleRef: &api.ObjectRef{
-			Name: postRuleName,
-			Uid:  postRuleUid,
-		},
-	}
-	_, err := backupDriver.CreateBackup(ctx, bkpCreateRequest)
-	log.FailOnError(err, "Failed to take backup with request -\n%v", bkpCreateRequest)
-}
-
-// ShareBackup provides access to the mentioned groups or/add users
-func ShareBackup(backupName string, groupNames []string, userNames []string, accessLevel BackupAccess, ctx context.Context) error {
-	var bkpUid string
-	backupDriver := Inst().Backup
-	groupIDs := make([]string, 0)
-	userIDs := make([]string, 0)
-
-	bkpUid, err := backupDriver.GetBackupUID(ctx, backupName, orgID)
-	log.FailOnError(err, "Failed to get backup UID for [%s]", backupName)
-	log.Infof("Backup UID for %s - %s", backupName, bkpUid)
-
-	for _, groupName := range groupNames {
-		groupID, err := backup.FetchIDOfGroup(groupName)
-		log.FailOnError(err, "Error fetching group ID")
-		groupIDs = append(groupIDs, groupID)
-	}
-
-	for _, userName := range userNames {
-		userID, err := backup.FetchIDOfUser(userName)
-		log.FailOnError(err, "Error fetching user ID")
-		userIDs = append(userIDs, userID)
-	}
-
-	groupBackupShareAccessConfigs := make([]*api.BackupShare_AccessConfig, 0)
-
-	for _, groupName := range groupNames {
-		groupBackupShareAccessConfig := &api.BackupShare_AccessConfig{
-			Id:     groupName,
-			Access: api.BackupShare_AccessType(accessLevel),
-		}
-		groupBackupShareAccessConfigs = append(groupBackupShareAccessConfigs, groupBackupShareAccessConfig)
-	}
-
-	userBackupShareAccessConfigs := make([]*api.BackupShare_AccessConfig, 0)
-
-	for _, userID := range userIDs {
-		userBackupShareAccessConfig := &api.BackupShare_AccessConfig{
-			Id:     userID,
-			Access: api.BackupShare_AccessType(accessLevel),
-		}
-		userBackupShareAccessConfigs = append(userBackupShareAccessConfigs, userBackupShareAccessConfig)
-	}
-
-	shareBackupRequest := &api.BackupShareUpdateRequest{
-		OrgId: orgID,
-		Name:  backupName,
-		Backupshare: &api.BackupShare{
-			Groups:        groupBackupShareAccessConfigs,
-			Collaborators: userBackupShareAccessConfigs,
-		},
-		Uid: bkpUid,
-	}
-
-	_, err = backupDriver.UpdateBackupShare(ctx, shareBackupRequest)
-	return err
-
-}
-
-// ClusterUpdateBackupShare shares all backup with the users and/or groups provided for a given cluster
-// addUsersOrGroups - provide true if the mentioned users/groups needs to be added
-// addUsersOrGroups - provide false if the mentioned users/groups needs to be deleted or removed
-func ClusterUpdateBackupShare(clusterName string, groupNames []string, userNames []string, accessLevel BackupAccess, addUsersOrGroups bool, ctx context.Context) error {
-	backupDriver := Inst().Backup
-	groupIDs := make([]string, 0)
-	userIDs := make([]string, 0)
-	_, clusterUID := backupDriver.RegisterBackupCluster(orgID, SourceClusterName, "")
-
-	for _, groupName := range groupNames {
-		groupID, err := backup.FetchIDOfGroup(groupName)
-		log.FailOnError(err, "Error fetching group ID")
-		groupIDs = append(groupIDs, groupID)
-	}
-
-	for _, userName := range userNames {
-		userID, err := backup.FetchIDOfUser(userName)
-		log.FailOnError(err, "Error fetching user ID")
-		userIDs = append(userIDs, userID)
-	}
-
-	groupBackupShareAccessConfigs := make([]*api.BackupShare_AccessConfig, 0)
-
-	for _, groupName := range groupNames {
-		groupBackupShareAccessConfig := &api.BackupShare_AccessConfig{
-			Id:     groupName,
-			Access: api.BackupShare_AccessType(accessLevel),
-		}
-		groupBackupShareAccessConfigs = append(groupBackupShareAccessConfigs, groupBackupShareAccessConfig)
-	}
-
-	userBackupShareAccessConfigs := make([]*api.BackupShare_AccessConfig, 0)
-
-	for _, userID := range userIDs {
-		userBackupShareAccessConfig := &api.BackupShare_AccessConfig{
-			Id:     userID,
-			Access: api.BackupShare_AccessType(accessLevel),
-		}
-		userBackupShareAccessConfigs = append(userBackupShareAccessConfigs, userBackupShareAccessConfig)
-	}
-
-	backupShare := &api.BackupShare{
-		Groups:        groupBackupShareAccessConfigs,
-		Collaborators: userBackupShareAccessConfigs,
-	}
-
-	var clusterBackupShareUpdateRequest *api.ClusterBackupShareUpdateRequest
-
-	if addUsersOrGroups {
-		clusterBackupShareUpdateRequest = &api.ClusterBackupShareUpdateRequest{
-			OrgId:          orgID,
-			Name:           clusterName,
-			AddBackupShare: backupShare,
-			DelBackupShare: nil,
-			Uid:            clusterUID,
-		}
-	} else {
-		clusterBackupShareUpdateRequest = &api.ClusterBackupShareUpdateRequest{
-			OrgId:          orgID,
-			Name:           clusterName,
-			AddBackupShare: nil,
-			DelBackupShare: backupShare,
-			Uid:            clusterUID,
-		}
-	}
-
-	_, err := backupDriver.ClusterUpdateBackupShare(ctx, clusterBackupShareUpdateRequest)
-	return err
-}
-
-func GetAllBackupsForUser(username, password string) ([]string, error) {
-	var bkp *api.BackupObject
-	backupNames := make([]string, 0)
-	backupDriver := Inst().Backup
-	ctx, err := backup.GetNonAdminCtx(username, password)
-	log.FailOnError(err, "Fetching %s ctx", username)
-
-	bkpEnumerateReq := &api.BackupEnumerateRequest{
-		OrgId: orgID}
-	curBackups, err := backupDriver.EnumerateBackup(ctx, bkpEnumerateReq)
-	for _, bkp = range curBackups.GetBackups() {
-		backupNames = append(backupNames, bkp.GetName())
-	}
-	return backupNames, err
-}
-
-func GetNodesForBackup(backupName string, bkpNamespace string,
-	orgID string, clusterName string, triggerOpts *driver_api.TriggerOptions) []node.Node {
-
-	var nodes []node.Node
-	backupDriver := Inst().Backup
-
-	backupUID := getBackupUID(backupName, orgID)
-	backupInspectReq := &api.BackupInspectRequest{
-		Name:  backupName,
-		OrgId: orgID,
-		Uid:   backupUID,
-	}
-	err := Inst().Backup.WaitForBackupRunning(context.Background(), backupInspectReq, defaultTimeout, defaultRetryInterval)
-	Expect(err).NotTo(HaveOccurred(),
-		fmt.Sprintf("Failed to wait for backup [%s] to start. Error: [%v]",
-			backupName, err))
-
-	clusterInspectReq := &api.ClusterInspectRequest{
-		OrgId:          orgID,
-		Name:           clusterName,
-		IncludeSecrets: true,
-	}
-	//ctx, err := backup.GetPxCentralAdminCtx()
-	ctx, err := backup.GetAdminCtxFromSecret()
-	Expect(err).NotTo(HaveOccurred(),
-		fmt.Sprintf("Failed to fetch px-central-admin ctx: [%v]",
-			err))
-	clusterInspectRes, err := backupDriver.InspectCluster(ctx, clusterInspectReq)
-	Expect(err).NotTo(HaveOccurred(),
-		fmt.Sprintf("Failed to inspect cluster [%s] in org [%s]. Error: [%v]",
-			clusterName, orgID, err))
-	Expect(clusterInspectRes).NotTo(BeNil(),
-		"Got an empty response while inspecting cluster [%s] in org [%s]", clusterName, orgID)
-
-	cluster := clusterInspectRes.GetCluster()
-	volumeBackupIDs, err := backupDriver.GetVolumeBackupIDs(context.Background(),
-		backupName, bkpNamespace, cluster, orgID)
-	Expect(err).NotTo(HaveOccurred(),
-		fmt.Sprintf("Failed to get volume backup IDs for backup [%s] in org [%s]. Error: [%v]",
-			backupName, orgID, err))
-	Expect(len(volumeBackupIDs)).NotTo(Equal(0),
-		"Got empty list of volumeBackup IDs from backup driver")
-
-	for _, backupID := range volumeBackupIDs {
-		n, err := Inst().V.GetNodeForBackup(backupID)
-		Expect(err).NotTo(HaveOccurred(),
-			fmt.Sprintf("Failed to get node on which backup [%s] in running. Error: [%v]",
-				backupName, err))
-
-		log.Debugf("Volume backup [%s] is running on node [%s], node id: [%s]\n",
-			backupID, n.GetHostname(), n.GetId())
-		nodes = append(nodes, n)
-	}
-	return nodes
-}
-
-// CreateRestore creates restore
-func CreateRestore(restoreName string, backupName string, namespaceMapping map[string]string, clusterName string,
-	orgID string, ctx context.Context, storageClassMapping map[string]string) error {
-
-	var bkp *api.BackupObject
-	var bkpUid string
-	backupDriver := Inst().Backup
-	log.Infof("Getting the UID of the backup needed to be restored")
-	bkpEnumerateReq := &api.BackupEnumerateRequest{
-		OrgId: orgID}
-	curBackups, err := backupDriver.EnumerateBackup(ctx, bkpEnumerateReq)
-	if err != nil {
-		return err
-	}
-	for _, bkp = range curBackups.GetBackups() {
-		if bkp.Name == backupName {
-			bkpUid = bkp.Uid
-			break
-		}
-	}
-	createRestoreReq := &api.RestoreCreateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  restoreName,
-			OrgId: orgID,
-		},
-		Backup:              backupName,
-		Cluster:             clusterName,
-		NamespaceMapping:    namespaceMapping,
-		StorageClassMapping: storageClassMapping,
-		BackupRef: &api.ObjectRef{
-			Name: backupName,
-			Uid:  bkpUid,
-		},
-	}
-	_, err = backupDriver.CreateRestore(ctx, createRestoreReq)
-	if err != nil {
-		return err
-	}
-	restoreInspectRequest := &api.RestoreInspectRequest{
-		Name:  restoreName,
-		OrgId: orgID,
-	}
-	restoreSuccessCheck := func() (interface{}, bool, error) {
-		resp, err := Inst().Backup.InspectRestore(ctx, restoreInspectRequest)
-		restoreResponseStatus := resp.GetRestore().GetStatus()
-		log.FailOnError(err, "Failed verifying restore for - %s", restoreName)
-		if restoreResponseStatus.GetStatus() == api.RestoreInfo_StatusInfo_PartialSuccess || restoreResponseStatus.GetStatus() == api.RestoreInfo_StatusInfo_Success {
-			log.Infof("Restore status - %s", restoreResponseStatus)
-			log.InfoD("Status of %s - [%s]",
-				restoreName, restoreResponseStatus.GetStatus())
-			return "", false, nil
-		}
-		return "", true, fmt.Errorf("expected status of %s - [%s] or [%s], but got [%s]",
-			restoreName, api.RestoreInfo_StatusInfo_PartialSuccess.String(), api.RestoreInfo_StatusInfo_Success, restoreResponseStatus.GetStatus())
-	}
-	task.DoRetryWithTimeout(restoreSuccessCheck, 10*time.Minute, 30*time.Second)
-	return nil
-}
-
-// CreateRestore creates restore
-func CreateRestoreWithUID(restoreName string, backupName string, namespaceMapping map[string]string, clusterName string,
-	orgID string, ctx context.Context, storageClassMapping map[string]string, backupUID string) error {
-
-	backupDriver := Inst().Backup
-	log.Infof("Getting the UID of the backup needed to be restored")
-
-	createRestoreReq := &api.RestoreCreateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  restoreName,
-			OrgId: orgID,
-		},
-		Backup:              backupName,
-		Cluster:             clusterName,
-		NamespaceMapping:    namespaceMapping,
-		StorageClassMapping: storageClassMapping,
-		BackupRef: &api.ObjectRef{
-			Name: backupName,
-			Uid:  backupUID,
-		},
-	}
-	_, err := backupDriver.CreateRestore(ctx, createRestoreReq)
-	if err != nil {
-		return err
-	}
-	restoreInspectRequest := &api.RestoreInspectRequest{
-		Name:  restoreName,
-		OrgId: orgID,
-	}
-	restoreSuccessCheck := func() (interface{}, bool, error) {
-		resp, err := Inst().Backup.InspectRestore(ctx, restoreInspectRequest)
-		if err != nil {
-			return "", false, err
-		}
-		restoreResponseStatus := resp.GetRestore().GetStatus()
-		if restoreResponseStatus.GetStatus() == api.RestoreInfo_StatusInfo_PartialSuccess || restoreResponseStatus.GetStatus() == api.RestoreInfo_StatusInfo_Success {
-			log.Infof("Restore status - %s", restoreResponseStatus)
-			log.InfoD("Status of %s - [%s]",
-				restoreName, restoreResponseStatus.GetStatus())
-			return "", false, nil
-		}
-		return "", true, fmt.Errorf("expected status of %s - [%s] or [%s], but got [%s]",
-			restoreName, api.RestoreInfo_StatusInfo_PartialSuccess.String(), api.RestoreInfo_StatusInfo_Success, restoreResponseStatus.GetStatus())
-	}
-	task.DoRetryWithTimeout(restoreSuccessCheck, 10*time.Minute, 30*time.Second)
-	return nil
-}
-
-// TearDownBackupRestoreSpecific deletes backups and restores specified by name as well as backup location
-func TearDownBackupRestoreSpecific(backups []string, restores []string) {
-	for _, backupName := range backups {
-		backupUID := getBackupUID(backupName, OrgID)
-		ctx, err := backup.GetAdminCtxFromSecret()
-		log.FailOnError(err, "Fetching px-central-admin ctx")
-		_, err = DeleteBackup(backupName, backupUID, OrgID, ctx)
-		dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting backup - [%s]", backupName))
-	}
-	for _, restoreName := range restores {
-		ctx, err := backup.GetAdminCtxFromSecret()
-		log.FailOnError(err, "Fetching px-central-admin ctx")
-		err = DeleteRestore(restoreName, OrgID, ctx)
-		dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting Restore %s", restoreName))
-	}
-	provider := GetProvider()
-	ctx, err := backup.GetAdminCtxFromSecret()
-	log.FailOnError(err, "Fetching px-central-admin ctx")
-	DeleteCluster(destinationClusterName, OrgID, ctx)
-	DeleteCluster(SourceClusterName, orgID, ctx)
-	// Need to add backup location UID for Delete Backup Location call
-	err = DeleteBackupLocation(backupLocationName, "", OrgID)
-	dash.VerifySafely(err, nil, fmt.Sprintf("Deleting backup location %s", backupLocationName))
-	DeleteCloudCredential(CredName, OrgID, CloudCredUID)
-	DeleteBucket(provider, BucketName)
-}
-
-// CreateRestoreWithoutCheck creates restore without waiting for completion
-func CreateRestoreWithoutCheck(restoreName string, backupName string,
-	namespaceMapping map[string]string, clusterName string, orgID string, ctx context.Context) error {
-
-	var bkp *api.BackupObject
-	var bkpUid string
-	backupDriver := Inst().Backup
-	log.Infof("Getting the UID of the backup needed to be restored")
-	bkpEnumerateReq := &api.BackupEnumerateRequest{
-		OrgId: orgID}
-	curBackups, _ := backupDriver.EnumerateBackup(ctx, bkpEnumerateReq)
-	log.Debugf("Enumerate backup response -\n%v", curBackups)
-	for _, bkp = range curBackups.GetBackups() {
-		if bkp.Name == backupName {
-			bkpUid = bkp.Uid
-			break
-		}
-	}
-	createRestoreReq := &api.RestoreCreateRequest{
-		CreateMetadata: &api.CreateMetadata{
-			Name:  restoreName,
-			OrgId: orgID,
-		},
-		Backup:           backupName,
-		Cluster:          clusterName,
-		NamespaceMapping: namespaceMapping,
-		BackupRef: &api.ObjectRef{
-			Name: backupName,
-			Uid:  bkpUid,
-		},
-	}
-	_, err := backupDriver.CreateRestore(ctx, createRestoreReq)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// CreateRestoreGetErr creates restore
-func CreateRestoreGetErr(restoreName string, backupName string,
-	namespaceMapping map[string]string, clusterName string, orgID string) (err error) {
-
-	Step(fmt.Sprintf("Create restore [%s] in org [%s] on cluster [%s]",
-		restoreName, orgID, clusterName), func() {
-
-		backupDriver := Inst().Backup
-		createRestoreReq := &api.RestoreCreateRequest{
-			CreateMetadata: &api.CreateMetadata{
-				Name:  restoreName,
-				OrgId: orgID,
-			},
-			Backup:           backupName,
-			Cluster:          clusterName,
-			NamespaceMapping: namespaceMapping,
-		}
-		//ctx, err := backup.GetPxCentralAdminCtx()
-		ctx, err := backup.GetAdminCtxFromSecret()
-		Expect(err).NotTo(HaveOccurred(),
-			fmt.Sprintf("Failed to fetch px-central-admin ctx: [%v]",
-				err))
-		_, err = backupDriver.CreateRestore(ctx, createRestoreReq)
-		if err != nil {
-			log.Errorf("Failed to create restore [%s] in org [%s] on cluster [%s]. Error: [%v]",
-				restoreName, orgID, clusterName, err)
-		}
-
-		// TODO: validate createClusterResponse also
-	})
-	return err
-}
-
-func getBackupUID(backupName, orgID string) string {
-	backupDriver := Inst().Backup
-	ctx, err := backup.GetAdminCtxFromSecret()
-	Expect(err).NotTo(HaveOccurred(),
-		fmt.Sprintf("Failed to fetch px-central-admin ctx: [%v]",
-			err))
-	backupUID, err := backupDriver.GetBackupUID(ctx, backupName, orgID)
-	Expect(err).NotTo(HaveOccurred(),
-		fmt.Sprintf("Failed to get backup uid for org %s backup %s ctx: [%v]",
-			orgID, backupName, err))
-	return backupUID
-}
-
-func getSizeOfMountPoint(podname string, namespace string, kubeconfigfile string) (int, error) {
-	var number int
-	ret, err := kubectlExec([]string{podname, "-n", namespace, "--kubeconfig=", kubeconfigfile, " -- /bin/df"})
-	if err != nil {
-		return 0, err
-	}
-	for _, line := range strings.SplitAfter(ret, "\n") {
-		if strings.Contains(line, "pxd") {
-			ret = strings.Fields(line)[3]
-		}
-	}
-	number, err = strconv.Atoi(ret)
-	if err != nil {
-		return 0, err
-	}
-	return number, nil
-}
-
-func kubectlExec(arguments []string) (string, error) {
-	if len(arguments) == 0 {
-		return "", fmt.Errorf("no arguments supplied for kubectl command")
-	}
-	cmd := exec.Command("kubectl exec -it", arguments...)
-	output, err := cmd.Output()
-	log.Debugf("command output for '%s': %s", cmd.String(), string(output))
-	if err != nil {
-		return "", fmt.Errorf("error on executing kubectl command, Err: %+v", err)
-	}
-	return string(output), err
-}
-
-func createUsers(numberOfUsers int) []string {
-	users := make([]string, 0)
-	log.InfoD("Creating %d users", numberOfUsers)
-	var wg sync.WaitGroup
-	for i := 1; i <= numberOfUsers; i++ {
-		userName := fmt.Sprintf("testuser%v-%v", i, time.Now().Unix())
-		firstName := fmt.Sprintf("FirstName%v", i)
-		lastName := fmt.Sprintf("LastName%v", i)
-		email := fmt.Sprintf("%v@cnbu.com", userName)
-		wg.Add(1)
-		go func(userName, firstName, lastName, email string) {
-			err := backup.AddUser(userName, firstName, lastName, email, "Password1")
-			log.FailOnError(err, "Failed to create user - %s", userName)
-			users = append(users, userName)
-			wg.Done()
-		}(userName, firstName, lastName, email)
-	}
-	wg.Wait()
-	return users
-}
-
-func DeleteCloudAccounts(backupLocationMap map[string]string, credName string, cloudCredUID string, ctx context.Context) {
-	if len(backupLocationMap) != 0 {
-		for backupLocationUID, bkpLocationName := range backupLocationMap {
-			err := DeleteBackupLocation(bkpLocationName, backupLocationUID, orgID)
-			dash.VerifySafely(err, nil, fmt.Sprintf("Deleting backup location %s", bkpLocationName))
-		}
-		time.Sleep(time.Minute * 3)
-		DeleteCloudCredential(credName, orgID, cloudCredUID)
-
-	}
-	DeleteCluster(SourceClusterName, orgID, ctx)
-	DeleteCluster(destinationClusterName, orgID, ctx)
-}
-
-// AddRoleAndAccessToUsers assigns role and access level to the users
-// AddRoleAndAccessToUsers return map whose key is userRoleAccess and value is backup for each user
-func AddRoleAndAccessToUsers(users []string, backupNames []string) (map[userRoleAccess]string, error) {
-	var access BackupAccess
-	var role backup.PxBackupRole
-	roleAccessUserBackupContext := make(map[userRoleAccess]string)
-	ctx, err := backup.GetAdminCtxFromSecret()
-	log.FailOnError(err, "Fetching px-central-admin ctx")
-	for i := 0; i < len(users); i++ {
-		userIndex := i % 9
-		switch userIndex {
-		case 0:
-			access = ViewOnlyAccess
-			role = backup.ApplicationOwner
-		case 1:
-			access = RestoreAccess
-			role = backup.ApplicationOwner
-		case 2:
-			access = FullAccess
-			role = backup.ApplicationOwner
-		case 3:
-			access = ViewOnlyAccess
-			role = backup.ApplicationUser
-		case 4:
-			access = RestoreAccess
-			role = backup.ApplicationUser
-		case 5:
-			access = FullAccess
-			role = backup.ApplicationUser
-		case 6:
-			access = ViewOnlyAccess
-			role = backup.InfrastructureOwner
-		case 7:
-			access = RestoreAccess
-			role = backup.InfrastructureOwner
-		case 8:
-			access = FullAccess
-			role = backup.InfrastructureOwner
-		default:
-			access = ViewOnlyAccess
-			role = backup.ApplicationOwner
-		}
-		ctxNonAdmin, err := backup.GetNonAdminCtx(users[i], "Password1")
-		log.FailOnError(err, "Fetching user ctx")
-		userRoleAccessContext := userRoleAccess{users[i], role, access, ctxNonAdmin}
-		roleAccessUserBackupContext[userRoleAccessContext] = backupNames[i]
-		err = backup.AddRoleToUser(users[i], role, "Adding role to user")
-		if err != nil {
-			err = fmt.Errorf("Failed to add role %s to user %s with err %v", role, users[i], err)
-			return nil, err
-		}
-		err = ShareBackup(backupNames[i], nil, []string{users[i]}, access, ctx)
-		if err != nil {
-			return nil, err
-		}
-		log.Infof(" Backup %s shared with user %s", backupNames[i], users[i])
-	}
-	return roleAccessUserBackupContext, nil
-}
-func ValidateSharedBackupWithUsers(user string, access BackupAccess, backupName string, restoreName string) {
-	ctx, err := backup.GetAdminCtxFromSecret()
-	log.FailOnError(err, "Fetching px-central-admin ctx")
-	userCtx, err := backup.GetNonAdminCtx(user, "Password1")
-	log.FailOnError(err, fmt.Sprintf("Fetching %s user ctx", user))
-	log.InfoD("Registering Source and Destination clusters from user context")
-	CreateSourceAndDestClusters(orgID, "", "", userCtx)
-	log.InfoD("Validating if user [%s] with access [%v] can restore and delete backup %s or not", user, backupAccessKeyValue[access], backupName)
-	backupDriver := Inst().Backup
-	switch access {
-	case ViewOnlyAccess:
-		// Try restore with user having ViewOnlyAccess and it should fail
-		err := CreateRestore(restoreName, backupName, make(map[string]string), destinationClusterName, orgID, userCtx, make(map[string]string))
-		dash.VerifyFatal(strings.Contains(err.Error(), "failed to retrieve backup location"), true, "Verifying backup restore is not possible")
-		// Try to delete the backup with user having ViewOnlyAccess, and it should not pass
-		backupUID, err := backupDriver.GetBackupUID(ctx, backupName, orgID)
-		log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupName)
-		// Delete backup to confirm that the user has ViewOnlyAccess and cannot delete backup
-		_, err = DeleteBackup(backupName, backupUID, orgID, userCtx)
-		dash.VerifyFatal(strings.Contains(err.Error(), "doesn't have permission to delete backup"), true, "Verifying backup deletion is not possible")
-
-	case RestoreAccess:
-		// Try restore with user having RestoreAccess and it should pass
-		err := CreateRestore(restoreName, backupName, make(map[string]string), destinationClusterName, orgID, userCtx, make(map[string]string))
-		dash.VerifyFatal(err, nil, "Verifying that restore is possible")
-		// Try to delete the backup with user having RestoreAccess, and it should not pass
-		backupUID, err := backupDriver.GetBackupUID(ctx, backupName, orgID)
-		log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupName)
-		// Delete backup to confirm that the user has Restore Access and delete backup should fail
-		_, err = DeleteBackup(backupName, backupUID, orgID, userCtx)
-		dash.VerifyFatal(strings.Contains(err.Error(), "doesn't have permission to delete backup"), true, "Verifying backup deletion is not possible")
-
-	case FullAccess:
-		// Try restore with user having FullAccess, and it should pass
-		err := CreateRestore(restoreName, backupName, make(map[string]string), destinationClusterName, orgID, userCtx, make(map[string]string))
-		dash.VerifyFatal(err, nil, "Verifying that restore is possible")
-		// Try to delete the backup with user having FullAccess, and it should pass
-		backupUID, err := backupDriver.GetBackupUID(ctx, backupName, orgID)
-		log.FailOnError(err, "Failed while trying to get backup UID for - %s", backupName)
-		// Delete backup to confirm that the user has Full Access
-		_, err = DeleteBackup(backupName, backupUID, orgID, userCtx)
-		dash.VerifyFatal(err, nil, "Verifying that delete backup is possible")
-	}
-}
-
-func TearDownBackupRestore(bkpNamespaces []string, restoreNamespaces []string) {
-	for _, bkpNamespace := range bkpNamespaces {
-		BackupName := fmt.Sprintf("%s-%s", BackupNamePrefix, bkpNamespace)
-		backupUID := getBackupUID(BackupName, OrgID)
-		ctx, err := backup.GetAdminCtxFromSecret()
-		log.FailOnError(err, "Fetching px-central-admin ctx")
-		_, err = DeleteBackup(BackupName, backupUID, OrgID, ctx)
-		dash.VerifyFatal(err, nil, "Verifying that delete backup is possible")
-	}
-	for _, restoreNamespace := range restoreNamespaces {
-		ctx, err := backup.GetAdminCtxFromSecret()
-		log.FailOnError(err, "Fetching px-central-admin ctx")
-		RestoreName := fmt.Sprintf("%s-%s", restoreNamePrefix, restoreNamespace)
-		err = DeleteRestore(RestoreName, OrgID, ctx)
-		dash.VerifyFatal(err, nil, fmt.Sprintf("Deleting Restore %s", RestoreName))
-	}
-
-	ctx, err := backup.GetAdminCtxFromSecret()
-	log.FailOnError(err, "Fetching px-central-admin ctx")
-	DeleteCluster(destinationClusterName, OrgID, ctx)
-	DeleteCluster(SourceClusterName, orgID, ctx)
-	// Need to add backup location UID for Delete Backup Location call
-	err = DeleteBackupLocation(backupLocationName, "", OrgID)
-	dash.VerifySafely(err, nil, fmt.Sprintf("Deleting backup location %s", backupLocationName))
-	DeleteCloudCredential(CredName, OrgID, CloudCredUID)
-}
-
-func getEnv(environmentVariable string, defaultValue string) string {
-	value, present := os.LookupEnv(environmentVariable)
-	if !present {
-		value = defaultValue
-	}
-	return value
-}
-
-// ShareBackupWithUsersAndAccessAssignment shares backup with multiple users with different access levels
-// It returns a map with key as userAccessContext and value as backup shared
-func ShareBackupWithUsersAndAccessAssignment(backupNames []string, users []string, ctx context.Context) (map[userAccessContext]string, error) {
-	log.InfoD("Sharing backups with users with different access level")
-	accessUserBackupContext := make(map[userAccessContext]string)
-	var err error
-	var ctxNonAdmin context.Context
-	var access BackupAccess
-	for i, user := range users {
-		userIndex := i % 3
-		switch userIndex {
-		case 0:
-			access = ViewOnlyAccess
-			err = ShareBackup(backupNames[i], nil, []string{user}, access, ctx)
-		case 1:
-			access = RestoreAccess
-			err = ShareBackup(backupNames[i], nil, []string{user}, access, ctx)
-		case 2:
-			access = FullAccess
-			err = ShareBackup(backupNames[i], nil, []string{user}, access, ctx)
-		default:
-			access = ViewOnlyAccess
-			err = ShareBackup(backupNames[i], nil, []string{user}, access, ctx)
-		}
-		if err != nil {
-			return accessUserBackupContext, fmt.Errorf("unable to share backup %s with user %s Error: %v", backupNames[i], user, err)
-		}
-		ctxNonAdmin, err = backup.GetNonAdminCtx(users[i], "Password1")
-		if err != nil {
-			return accessUserBackupContext, fmt.Errorf("unable to get user context: %v", err)
-		}
-		accessContextUser := userAccessContext{users[i], access, ctxNonAdmin}
-		accessUserBackupContext[accessContextUser] = backupNames[i]
-	}
-	return accessUserBackupContext, nil
-}
-
-// GetAllBackupsAdmin returns all the backups that px-central-admin has access to
-func GetAllBackupsAdmin() ([]string, error) {
-	var bkp *api.BackupObject
-	backupNames := make([]string, 0)
-	backupDriver := Inst().Backup
-	ctx, err := backup.GetAdminCtxFromSecret()
-	log.FailOnError(err, "Fetching ctx")
-
-	bkpEnumerateReq := &api.BackupEnumerateRequest{
-		OrgId: orgID}
-	curBackups, err := backupDriver.EnumerateBackup(ctx, bkpEnumerateReq)
-	log.FailOnError(err, "Failed to enumerate on backup objects")
-	for _, bkp = range curBackups.GetBackups() {
-		backupNames = append(backupNames, bkp.GetName())
-	}
-	return backupNames, err
-}
-
-// GetAllRestoresAdmin returns all the backups that px-central-admin has access to
-func GetAllRestoresAdmin() ([]string, error) {
-	restoreNames := make([]string, 0)
-	backupDriver := Inst().Backup
-	ctx, err := backup.GetAdminCtxFromSecret()
-	if err != nil {
-		return restoreNames, err
-	}
-
-	restoreEnumerateRequest := &api.RestoreEnumerateRequest{
-		OrgId: orgID,
-	}
-	restoreResponse, err := backupDriver.EnumerateRestore(ctx, restoreEnumerateRequest)
-	if err != nil {
-		return restoreNames, err
-	}
-	for _, restore := range restoreResponse.GetRestores() {
-		restoreNames = append(restoreNames, restore.Name)
-	}
-	return restoreNames, nil
-}
-
-// TODO: There is no delete org API
-/*func DeleteOrganization(orgID string) {
-	Step(fmt.Sprintf("Delete organization [%s]", orgID), func() {
-		backupDriver := Inst().Backup
-		req := &api.Delete{
-			CreateMetadata: &api.CreateMetadata{
-				Name: orgID,
-			},
-		}
-		_, err := backupDriver.Delete(req)
-		Expect(err).NotTo(HaveOccurred())
-	})
-}*/
-
-func generateEncryptionKey() string {
-	var lower = []byte("abcdefghijklmnopqrstuvwxyz")
-	var upper = []byte("ABCDEFGHIJKLMNOPQRSTUVWXYZ")
-	var number = []byte("0123456789")
-	var special = []byte("~=+%^*/()[]{}/!@#$?|")
-	allChar := append(lower, upper...)
-	allChar = append(allChar, number...)
-	allChar = append(allChar, special...)
-
-	b := make([]byte, 12)
-	// select 1 upper, 1 lower, 1 number and 1 special
-	b[0] = lower[rand.Intn(len(lower))]
-	b[1] = upper[rand.Intn(len(upper))]
-	b[2] = number[rand.Intn(len(number))]
-	b[3] = special[rand.Intn(len(special))]
-	for i := 4; i < 12; i++ {
-		// randomly select 1 character from given charset
-		b[i] = allChar[rand.Intn(len(allChar))]
-	}
-
-	//shuffle character
-	rand.Shuffle(len(b), func(i, j int) {
-		b[i], b[j] = b[j], b[i]
-	})
-
-	return string(b)
-}
-
-func GetScheduleUID(scheduleName string, orgID string, ctx context.Context) (string, error) {
-	backupDriver := Inst().Backup
-	bkpUid, err := backupDriver.GetBackupUID(ctx, scheduleName, orgID)
-	if err != nil {
-		return "", fmt.Errorf("failed to get backup UID for %s, Err: %v", scheduleName, err)
-	}
-	backupSchInspectRequest := &api.BackupScheduleInspectRequest{
-		Name:  scheduleName,
-		Uid:   bkpUid,
-		OrgId: orgID,
-	}
-	resp, err := backupDriver.InspectBackupSchedule(ctx, backupSchInspectRequest)
-	if err != nil {
-		return "", fmt.Errorf("failed to inspect backup [%s] with UID [%s], Err: %v", scheduleName, bkpUid, err)
-	}
-	scheduleUid := resp.GetBackupSchedule().GetUid()
-	log.InfoD("Schedule Name - %s Schedule UID - %s", scheduleName, scheduleUid)
-
-	return scheduleUid, err
-}
-
-func removeStringItemFromSlice(itemList []string, item []string) []string {
-	sort.Sort(sort.StringSlice(itemList))
-	for _, element := range item {
-		index := sort.StringSlice(itemList).Search(element)
-		itemList = append(itemList[:index], itemList[index+1:]...)
-	}
-	return itemList
-}
-
-func removeIntItemFromSlice(itemList []int, item []int) []int {
-	sort.Sort(sort.IntSlice(itemList))
-	for _, element := range item {
-		index := sort.IntSlice(itemList).Search(element)
-		itemList = append(itemList[:index], itemList[index+1:]...)
-	}
-	return itemList
-}
-
-func getAllBackupLocations(ctx context.Context) (map[string]string, error) {
-	backupLocationMap := make(map[string]string, 0)
-	backupDriver := Inst().Backup
-	backupLocationEnumerateRequest := &api.BackupLocationEnumerateRequest{
-		OrgId: orgID,
-	}
-	response, err := backupDriver.EnumerateBackupLocation(ctx, backupLocationEnumerateRequest)
-	if err != nil {
-		return backupLocationMap, err
-	}
-	if len(response.BackupLocations) > 0 {
-		log.Infof("Found backup locations - ")
-		for _, backupLocation := range response.BackupLocations {
-			backupLocationMap[backupLocation.Uid] = backupLocation.Name
-			log.Infof(backupLocation.Name)
-		}
-	} else {
-		log.Info("No backup locations found")
-	}
-	return backupLocationMap, nil
-}
-
-func getAllCloudCredentials(ctx context.Context) (map[string]string, error) {
-	cloudCredentialMap := make(map[string]string, 0)
-	backupDriver := Inst().Backup
-	cloudCredentialEnumerateRequest := &api.CloudCredentialEnumerateRequest{
-		OrgId: orgID,
-	}
-	response, err := backupDriver.EnumerateCloudCredential(ctx, cloudCredentialEnumerateRequest)
-	if err != nil {
-		return cloudCredentialMap, err
-	}
-	if len(response.CloudCredentials) > 0 {
-		log.Infof("Found cloud credentials - ")
-		for _, cloudCredential := range response.CloudCredentials {
-			cloudCredentialMap[cloudCredential.Uid] = cloudCredential.Name
-			log.Infof(cloudCredential.Name)
-		}
-	} else {
-		log.Info("No cloud credentials found")
-	}
-	return cloudCredentialMap, nil
-}
-
-func GetAllRestoresNonAdminCtx(ctx context.Context) ([]string, error) {
-	restoreNames := make([]string, 0)
-	backupDriver := Inst().Backup
-	restoreEnumerateRequest := &api.RestoreEnumerateRequest{
-		OrgId: orgID,
-	}
-	restoreResponse, err := backupDriver.EnumerateRestore(ctx, restoreEnumerateRequest)
-	if err != nil {
-		return restoreNames, err
-	}
-	for _, restore := range restoreResponse.GetRestores() {
-		restoreNames = append(restoreNames, restore.Name)
-	}
-	return restoreNames, nil
-}
