+++ tests/common.go
@@ -1975,26 +1975,21 @@ func updateNamespace(in interface{}, namespaceMapping map[string]string) error {
 }
 
 // DeleteCloudCredential deletes cloud credentials
-func DeleteCloudCredential(name string, orgID string, cloudCredUID string) {
-	Step(fmt.Sprintf("Delete cloud credential [%s] in org [%s]", name, orgID), func() {
-		backupDriver := Inst().Backup
+func DeleteCloudCredential(name string, orgID string, cloudCredUID string) error {
 
-		credDeleteRequest := &api.CloudCredentialDeleteRequest{
-			Name:  name,
-			OrgId: orgID,
-			Uid:   cloudCredUID,
-		}
-		//ctx, err := backup.GetPxCentralAdminCtx()
-		ctx, err := backup.GetAdminCtxFromSecret()
-		expect(err).NotTo(haveOccurred(),
-			fmt.Sprintf("Failed to fetch px-central-admin ctx: [%v]",
-				err))
-		backupDriver.DeleteCloudCredential(ctx, credDeleteRequest)
-		// Best effort cleanup, dont fail test, if deletion fails
-		// expect(err).NotTo(haveOccurred(),
-		//  fmt.Sprintf("Failed to delete cloud credential [%s] in org [%s]", name, orgID))
-		// TODO: validate CreateCloudCredentialResponse also
-	})
+	backupDriver := Inst().Backup
+	credDeleteRequest := &api.CloudCredentialDeleteRequest{
+		Name:  name,
+		OrgId: orgID,
+		Uid:   cloudCredUID,
+	}
+	//ctx, err := backup.GetPxCentralAdminCtx()
+	ctx, err := backup.GetAdminCtxFromSecret()
+	if err != nil {
+		return err
+	}
+	_, err = backupDriver.DeleteCloudCredential(ctx, credDeleteRequest)
+	return err
 }
 
 // ValidateVolumeParametersGetErr validates volume parameters using volume driver and returns err instead of failing
@@ -2734,23 +2729,19 @@ func DeleteBackup(backupName string, backupUID string, orgID string, ctx context
 }
 
 // DeleteCluster deletes/de-registers cluster from px-backup
-func DeleteCluster(name string, orgID string, ctx context1.Context) {
+func DeleteCluster(name string, orgID string, ctx context1.Context) error {
 
-	Step(fmt.Sprintf("Delete cluster [%s] in org [%s]", name, orgID), func() {
-		backupDriver := Inst().Backup
-		clusterDeleteReq := &api.ClusterDeleteRequest{
-			OrgId: orgID,
-			Name:  name,
-		}
-		ctx, err := backup.GetPxCentralAdminCtx()
-		expect(err).NotTo(haveOccurred(),
-			fmt.Sprintf("Failed to fetch px-central-admin ctx: [%v]",
-				err))
-		backupDriver.DeleteCluster(ctx, clusterDeleteReq)
-		// Best effort cleanup, dont fail test, if deletion fails
-		//expect(err).NotTo(haveOccurred(),
-		//	fmt.Sprintf("Failed to delete cluster [%s] in org [%s]", name, orgID))
-	})
+	backupDriver := Inst().Backup
+	clusterDeleteReq := &api.ClusterDeleteRequest{
+		OrgId: orgID,
+		Name:  name,
+	}
+	ctx, err := backup.GetPxCentralAdminCtx()
+	if err != nil {
+		return err
+	}
+	_, err = backupDriver.DeleteCluster(ctx, clusterDeleteReq)
+	return err
 }
 
 // DeleteBackupLocation deletes backup location
@@ -2828,7 +2819,7 @@ func DeleteSchedule(backupScheduleName, backupScheduleUID, schedulePolicyName, s
 // CreateSourceAndDestClusters creates source and destination cluster
 // 1st cluster in KUBECONFIGS ENV var is source cluster while
 // 2nd cluster is destination cluster
-func CreateSourceAndDestClusters(orgID string, cloud_name string, uid string, ctx context1.Context) error {
+func CreateSourceAndDestClusters(orgID string, cloudName string, uid string, ctx context1.Context) error {
 	// TODO: Add support for adding multiple clusters from
 	// comma separated list of kubeconfig files
 	kubeconfigs := os.Getenv("KUBECONFIGS")
@@ -2849,7 +2840,7 @@ func CreateSourceAndDestClusters(orgID string, cloud_name string, uid string, ct
 		return err
 	}
 	log.Infof("Save cluster %s kubeconfig to %s", SourceClusterName, srcClusterConfigPath)
-	err = CreateCluster(SourceClusterName, srcClusterConfigPath, orgID, cloud_name, uid, ctx)
+	err = CreateCluster(SourceClusterName, srcClusterConfigPath, orgID, cloudName, uid, ctx)
 	if err != nil {
 		return err
 	}
@@ -2860,7 +2851,8 @@ func CreateSourceAndDestClusters(orgID string, cloud_name string, uid string, ct
 		return err
 	}
 	log.Infof("Save cluster %s kubeconfig to %s", destinationClusterName, dstClusterConfigPath)
-	err = CreateCluster(destinationClusterName, dstClusterConfigPath, orgID, cloud_name, uid, ctx)
+
+	err = CreateCluster(destinationClusterName, dstClusterConfigPath, orgID, cloudName, uid, ctx)
 	if err != nil {
 		return err
 	}
@@ -3807,11 +3799,14 @@ func TearDownBackupRestoreAll() {
 			RetrySeconds*time.Second)
 	}
 	provider := GetProvider()
-	DeleteCluster(destinationClusterName, OrgID, ctx)
-	DeleteCluster(SourceClusterName, OrgID, ctx)
+	err = DeleteCluster(destinationClusterName, OrgID, ctx)
+	log.FailOnError(err, fmt.Sprintf("Deleting cluster %s", destinationClusterName))
+	err = DeleteCluster(SourceClusterName, OrgID, ctx)
+	log.FailOnError(err, fmt.Sprintf("Deleting cluster %s", SourceClusterName))
 	err = DeleteBackupLocation(backupLocationName, BackupLocationUID, OrgID)
 	dash.VerifySafely(err, nil, fmt.Sprintf("Deleting backup location %s", backupLocationName))
-	DeleteCloudCredential(CredName, OrgID, CloudCredUID)
+	err = DeleteCloudCredential(CredName, OrgID, CloudCredUID)
+	dash.VerifySafely(err, nil, fmt.Sprintf("Deleting cloud cred %s", CredName))
 	DeleteBucket(provider, BucketName)
 }
